requirements.txt
vk_api.py
cahrdet
websockets
aiohttp
